QR Code Reader & Generator/requirement.txt
qrcode
pypng
PyQRCode
python-barcode